mejoras

+++ routes/web.php
@@ -30,59 +30,59 @@ Route::get('/symlink', function () {
 Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
 
 
-Route::resource('juzgados', App\Http\Controllers\JuzgadoController::class);
+Route::resource('juzgados', App\Http\Controllers\JuzgadoController::class)->middleware('auth');
 
 
-Route::resource('circunscripcions', App\Http\Controllers\CircunscripcionController::class);
+Route::resource('circunscripcions', App\Http\Controllers\CircunscripcionController::class)->middleware('auth');
 
 
-Route::resource('circunscripcionJuzgados', App\Http\Controllers\Circunscripcion_juzgadoController::class);
+Route::resource('circunscripcionJuzgados', App\Http\Controllers\Circunscripcion_juzgadoController::class)->middleware('auth');
 
 
-Route::resource('clientes', App\Http\Controllers\ClienteController::class);
+Route::resource('clientes', App\Http\Controllers\ClienteController::class)->middleware('auth');
 
 
-Route::resource('expedientes', App\Http\Controllers\ExpedienteController::class);
+Route::resource('expedientes', App\Http\Controllers\ExpedienteController::class)->middleware('auth');
 
 
-Route::resource('gastoExpedientes', App\Http\Controllers\Gasto_expedienteController::class);
-
-
-
-Route::post('/clientes/crear', [App\Http\Controllers\ClienteController::class, 'crear']);
-Route::post('/juzgados/crear', [App\Http\Controllers\JuzgadoController::class, 'crear']);
-Route::post('/circunscripcions/crear', [App\Http\Controllers\CircunscripcionController::class, 'crear']);
-
-Route::resource('expedienteDetalles', App\Http\Controllers\Expediente_detalleController::class);
-
-
-Route::resource('pagoExpedientes', App\Http\Controllers\Pago_expedienteController::class);
-Route::get('pagoExpediente/archivo', [App\Http\Controllers\Pago_expedienteController::class, 'archivo'])->name('pagoExpedientes.archivo');
-Route::get('gastoExpediente/archivo', [App\Http\Controllers\Gasto_expedienteController::class, 'archivo'])->name('gastoExpedientes.archivo');
-
-
-Route::resource('users', App\Http\Controllers\UserController::class);
+Route::resource('gastoExpedientes', App\Http\Controllers\Gasto_expedienteController::class)->middleware('auth');
 
 
 
-Route::get('audiencias', [App\Http\Controllers\AudienciasController::class, 'index'])->name('audiencias.index');
-Route::post('audiencias', [App\Http\Controllers\AudienciasController::class, 'store'])->name('audiencias.store');
-Route::patch('audiencias/update/{id}', [App\Http\Controllers\AudienciasController::class, 'update'])->name('audiencias.update');
-Route::delete('audiencias/destroy/{id}', [App\Http\Controllers\AudienciasController::class, 'destroy'])->name('audiencias.destroy');
+Route::post('/clientes/crear', [App\Http\Controllers\ClienteController::class, 'crear'])->middleware('auth');
+Route::post('/juzgados/crear', [App\Http\Controllers\JuzgadoController::class, 'crear'])->middleware('auth');
+Route::post('/circunscripcions/crear', [App\Http\Controllers\CircunscripcionController::class, 'crear'])->middleware('auth');
 
-Route::get('/audiencias/mostrar', [App\Http\Controllers\AudienciasController::class, 'mostrar']);
-Route::post('audiencia_crear', [App\Http\Controllers\AudienciasController::class, 'audiencia']);	
-
-Route::resource('reporte', App\Http\Controllers\ReporteController::class);	
-
-Route::resource('pdf', App\Http\Controllers\PdfController::class);
-
-Route::get('graficos/index', [App\Http\Controllers\GraficoController::class, 'grafico'])->name('graficos');
-Route::get('audiencias/consulta', [App\Http\Controllers\Consulta_audiencaController::class, 'consulta'])->name('consultas');
-
-Route::resource('fueros', App\Http\Controllers\FueroController::class);
+Route::resource('expedienteDetalles', App\Http\Controllers\Expediente_detalleController::class)->middleware('auth');
 
 
-Route::get('consultas', [App\Http\Controllers\ReporteController::class, 'consulta_expediente'])->name('consultas.expedientes');
+Route::resource('pagoExpedientes', App\Http\Controllers\Pago_expedienteController::class)->middleware('auth');
+Route::get('pagoExpediente/archivo', [App\Http\Controllers\Pago_expedienteController::class, 'archivo'])->name('pagoExpedientes.archivo')->middleware('auth');
+Route::get('gastoExpediente/archivo', [App\Http\Controllers\Gasto_expedienteController::class, 'archivo'])->name('gastoExpedientes.archivo')->middleware('auth');
 
-Route::resource('seguimintos', App\Http\Controllers\SeguimintoController::class);
+
+Route::resource('users', App\Http\Controllers\UserController::class)->middleware('auth');
+
+
+
+Route::get('audiencias', [App\Http\Controllers\AudienciasController::class, 'index'])->name('audiencias.index')->middleware('auth');
+Route::post('audiencias', [App\Http\Controllers\AudienciasController::class, 'store'])->name('audiencias.store')->middleware('auth');
+Route::patch('audiencias/update/{id}', [App\Http\Controllers\AudienciasController::class, 'update'])->name('audiencias.update')->middleware('auth');
+Route::delete('audiencias/destroy/{id}', [App\Http\Controllers\AudienciasController::class, 'destroy'])->name('audiencias.destroy')->middleware('auth');
+
+Route::get('/audiencias/mostrar', [App\Http\Controllers\AudienciasController::class, 'mostrar'])->middleware('auth');
+Route::post('audiencia_crear', [App\Http\Controllers\AudienciasController::class, 'audiencia'])->middleware('auth');	
+
+Route::resource('reporte', App\Http\Controllers\ReporteController::class)->middleware('auth');
+
+Route::resource('pdf', App\Http\Controllers\PdfController::class)->middleware('auth');
+
+Route::get('graficos/index', [App\Http\Controllers\GraficoController::class, 'grafico'])->name('graficos')->middleware('auth');
+Route::get('audiencias/consulta', [App\Http\Controllers\Consulta_audiencaController::class, 'consulta'])->name('consultas')->middleware('auth');
+
+Route::resource('fueros', App\Http\Controllers\FueroController::class)->middleware('auth');
+
+
+Route::get('consultas', [App\Http\Controllers\ReporteController::class, 'consulta_expediente'])->name('consultas.expedientes')->middleware('auth');
+
+Route::resource('seguimintos', App\Http\Controllers\SeguimintoController::class)->middleware('auth');
